+++ src/hooks/index.ts
@@ -7,3 +7,4 @@ export { useClickOutside } from './useClickOutside';
 export { usePrevious } from './usePrevious';
 export { useToggle } from './useToggle';
 export { useTimeout } from './useTimeout';
+export { useInterval } from './useInterval';